renames ExprContext.ts to utils.ts
