+++ src/features/expressions/index.test.ts
@@ -1,8 +1,8 @@
 import { NodeNotFoundWithoutContext } from 'src/features/expressions/errors';
 import { evalExpr } from 'src/features/expressions/index';
 import { ExprVal } from 'src/features/expressions/types';
-import type { ExpressionDataSources } from 'src/features/expressions/ExprContext';
 import type { ExprConfig } from 'src/features/expressions/types';
+import type { ExpressionDataSources } from 'src/features/expressions/utils';
 
 describe('Expressions', () => {
   it('should return default value if expression evaluates to null', () => {
